refactor: split watchdog into shelly/ and zero/ subfolders
