+++ watchdog/shelly/device/setup_shelly_watchdog.sh
