+++ watchdog/shelly/device/update_shelly_watchdog.sh
